"Create hook for using time zomes"
